Implement basic incognalio render function
The render function wraps React's native render function,
appending a container to the document body to use as a mount
point for a custom component. The component is then mounted,
given the dom node corresponding to the "selector" parameter
(our "anchor" node) as a prop, the idea being that it can then observe
changes on said node and react according to arbitrary specifications.

Note: The render function is (at this point) totally untested...